+++ custom_components/homematicip_local/switch.py
@@ -9,17 +9,14 @@ from aiohomematic.const import DataPointCategory
 from aiohomematic.model.custom import CustomDpSwitch
 from aiohomematic.model.generic import DpSwitch
 from aiohomematic.model.hub import ProgramDpSwitch, SysvarDpSwitch
-import voluptuous as vol
 
 from homeassistant.components.switch import SwitchEntity
 from homeassistant.const import STATE_ON, STATE_UNAVAILABLE, STATE_UNKNOWN
 from homeassistant.core import HomeAssistant, callback
-from homeassistant.helpers import entity_platform
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import HomematicConfigEntry
-from .const import HmipLocalServices
 from .control_unit import ControlUnit, signal_new_data_point
 from .generic_entity import (
     AioHomematicGenericProgramEntity,
@@ -28,7 +25,6 @@ from .generic_entity import (
 )
 
 _LOGGER = logging.getLogger(__name__)
-ATTR_ON_TIME: Final = "on_time"
 ATTR_CHANNEL_STATE: Final = "channel_state"
 
 
@@ -96,15 +92,6 @@ async def async_setup_entry(
 
     async_add_hub_switch(data_points=control_unit.get_new_hub_data_points(data_point_type=SysvarDpSwitch))
     # async_add_hub_switch(data_points=control_unit.get_new_hub_data_points(data_point_type=ProgramDpSwitch))
-
-    platform = entity_platform.async_get_current_platform()
-    platform.async_register_entity_service(
-        HmipLocalServices.SWITCH_SET_ON_TIME,
-        {
-            vol.Required(ATTR_ON_TIME): vol.All(vol.Coerce(int), vol.Range(min=-1, max=8580000)),
-        },
-        "async_set_on_time",
-    )
 
 
 class AioHomematicSwitch(AioHomematicGenericRestoreEntity[CustomDpSwitch | DpSwitch], SwitchEntity):